[alert_consumer] move KafkaSink to another file
